requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==36.0.1
Django==4.0.1
django-filter==21.1
django-rest-knox==4.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
review==1510
sqlparse==0.4.2
